Basic EditEventScreen setup
