+++ app/src/main/java/com/monkeyteam/chimpagne/ui/EventCreationScreen.kt
@@ -1,6 +1,5 @@
 package com.monkeyteam.chimpagne.ui
 
-import DateSelector
 import android.widget.Toast
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.layout.Arrangement
@@ -8,46 +7,27 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.rememberPagerState
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.CalendarToday
-import androidx.compose.material.icons.rounded.Description
-import androidx.compose.material.icons.rounded.LocationOn
-import androidx.compose.material.icons.rounded.Public
-import androidx.compose.material.icons.rounded.Tag
-import androidx.compose.material.icons.rounded.Title
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.unit.dp
 import com.monkeyteam.chimpagne.R
 import com.monkeyteam.chimpagne.ui.components.GoBackButton
-import com.monkeyteam.chimpagne.ui.components.Legend
-import com.monkeyteam.chimpagne.ui.components.LocationSelector
-import com.monkeyteam.chimpagne.ui.components.TagField
+import com.monkeyteam.chimpagne.ui.event.FirstPanel
+import com.monkeyteam.chimpagne.ui.event.FourthPanel
+import com.monkeyteam.chimpagne.ui.event.SecondPanel
+import com.monkeyteam.chimpagne.ui.event.ThirdPanel
 import com.monkeyteam.chimpagne.ui.navigation.NavigationActions
 import com.monkeyteam.chimpagne.viewmodels.EventViewModel
 import kotlinx.coroutines.launch
@@ -129,180 +109,5 @@ fun EventCreationScreen(
             }
       }
     }
-  }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun FirstPanel(eventViewModel: EventViewModel) {
-  val uiState by eventViewModel.uiState.collectAsState()
-
-  Column(modifier = Modifier.padding(16.dp)) {
-    Legend(
-        stringResource(id = R.string.event_creation_screen_title_legend),
-        Icons.Rounded.Title,
-        "Title")
-    Spacer(Modifier.height(16.dp))
-    OutlinedTextField(
-        value = uiState.title,
-        onValueChange = eventViewModel::updateEventTitle,
-        label = { Text(stringResource(id = R.string.event_creation_screen_title)) },
-        modifier = Modifier.fillMaxWidth().testTag("add_a_title"))
-
-    Spacer(modifier = Modifier.height(16.dp))
-
-    Legend(
-        stringResource(id = R.string.event_creation_screen_description_legend),
-        Icons.Rounded.Description,
-        "Description")
-
-    Spacer(modifier = Modifier.height(16.dp))
-
-    OutlinedTextField(
-        value = uiState.description,
-        onValueChange = eventViewModel::updateEventDescription,
-        label = { Text(stringResource(id = R.string.event_creation_screen_description)) },
-        modifier = Modifier.fillMaxWidth().testTag("add_a_description"),
-        maxLines = 3)
-    Spacer(modifier = Modifier.height(16.dp))
-
-    Legend(
-        stringResource(id = R.string.event_creation_screen_location_legend),
-        Icons.Rounded.LocationOn,
-        "Location")
-
-    Spacer(modifier = Modifier.height(16.dp))
-
-    LocationSelector(uiState.location, eventViewModel::updateEventLocation)
-
-    Spacer(modifier = Modifier.height(16.dp))
-
-    Legend(
-        stringResource(id = R.string.event_creation_screen_start_date_legend),
-        Icons.Rounded.CalendarToday,
-        "Start Date")
-
-    Spacer(modifier = Modifier.height(16.dp))
-
-    DateSelector(
-        selectedDate = uiState.startsAtCalendarDate,
-        onDateSelected = eventViewModel::updateEventStartCalendarDate,
-        modifier = Modifier.align(Alignment.CenterHorizontally),
-        selectTimeOfDay = true)
-
-    Spacer(modifier = Modifier.height(16.dp))
-
-    Legend(
-        stringResource(id = R.string.event_creation_screen_end_date_legend),
-        Icons.Rounded.CalendarToday,
-        "End Date")
-
-    Spacer(modifier = Modifier.height(16.dp))
-    // We will need to add some tests for DateSelector also
-    DateSelector(
-        selectedDate = uiState.endsAtCalendarDate,
-        onDateSelected = eventViewModel::updateEventEndCalendarDate,
-        modifier = Modifier.align(Alignment.CenterHorizontally),
-        selectTimeOfDay = true)
-  }
-}
-
-@Composable
-fun SecondPanel(eventViewModel: EventViewModel) {
-  val uiState by eventViewModel.uiState.collectAsState()
-
-  var tagFieldActive by remember { mutableStateOf(true) }
-
-  Column(modifier = Modifier.padding(16.dp)) {
-    Legend(
-        stringResource(id = R.string.event_creation_screen_tags_legend), Icons.Rounded.Tag, "Tags")
-
-    TagField(
-        uiState.tags,
-        eventViewModel::updateEventTags,
-        { tagFieldActive = it },
-        Modifier.fillMaxWidth().testTag("tag_field"))
-
-    Spacer(modifier = Modifier.height(16.dp))
-
-    Legend(
-        stringResource(id = R.string.event_creation_screen_public_legend),
-        Icons.Rounded.Public,
-        "Public")
-
-    Spacer(modifier = Modifier.height(16.dp))
-
-    Row(verticalAlignment = Alignment.CenterVertically) {
-      Checkbox(
-          checked = uiState.public,
-          onCheckedChange = { eventViewModel.updateEventPublicity(!uiState.public) })
-      if (uiState.public)
-          Text(stringResource(id = R.string.event_creation_screen_event_made_public))
-      else Text(stringResource(id = R.string.event_creation_screen_make_event_public))
-    }
-  }
-}
-
-@Composable
-fun ThirdPanel(eventViewModel: EventViewModel) {
-  val context = LocalContext.current
-  Column(modifier = Modifier.padding(16.dp)) {
-    Text(
-        stringResource(id = R.string.event_creation_screen_groceries),
-        style = MaterialTheme.typography.headlineSmall,
-        modifier = Modifier.testTag("groceries_title"))
-    Spacer(modifier = Modifier.height(16.dp))
-    Button(
-        onClick = {
-          Toast.makeText(
-                  context,
-                  context.getString(R.string.event_creation_screen_gorceries_toast),
-                  Toast.LENGTH_SHORT)
-              .show()
-        },
-        modifier = Modifier.testTag("add_groceries_button")) {
-          Text(stringResource(id = R.string.event_creation_screen_add_groceries))
-        }
-    Spacer(modifier = Modifier.height(16.dp))
-    LazyColumn {
-      // Populate with groceries items
-    }
-  }
-}
-
-// Comment to make a new commit
-@Composable
-fun FourthPanel(eventViewModel: EventViewModel) {
-
-  var parkingText by remember { mutableStateOf("") }
-  var bedsText by remember { mutableStateOf("") }
-  Column(modifier = Modifier.padding(16.dp)) {
-    Text(
-        stringResource(id = R.string.event_creation_screen_logistics),
-        style = MaterialTheme.typography.headlineSmall,
-        modifier = Modifier.testTag("logistics_title"))
-    Spacer(modifier = Modifier.height(16.dp))
-    Text(
-        stringResource(id = R.string.event_creation_screen_parking),
-        style = MaterialTheme.typography.bodyMedium,
-        modifier = Modifier.testTag("parking_title"))
-    OutlinedTextField(
-        value = parkingText,
-        onValueChange = { parkingText = it },
-        label = { Text(stringResource(id = R.string.event_creation_screen_number_parking)) },
-        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
-        modifier = Modifier.fillMaxWidth().testTag("n_parking"))
-
-    Spacer(modifier = Modifier.height(16.dp))
-    Text(
-        stringResource(id = R.string.event_creation_screen_beds),
-        style = MaterialTheme.typography.bodyMedium,
-        modifier = Modifier.testTag("beds_title"))
-    OutlinedTextField(
-        value = bedsText,
-        onValueChange = { bedsText = it },
-        label = { Text(stringResource(id = R.string.event_creation_screen_number_beds)) },
-        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
-        modifier = Modifier.fillMaxWidth().testTag("n_beds"))
   }
 }
