+++ app/src/main/res/values-fr/strings.xml
@@ -53,4 +53,6 @@
     <string name="event_creation_screen_gorceries_toast">Coming soon to chimpagne: groceries functionnality !</string>
     <string name="sign_in_with_google">Se connecter avec Google</string>
     <string name="welcome_screen_title">Bienvenue à</string>
+    <string name="save_changes">Sauvegarder les changements</string>
+    <string name="event_changes_saved">Les changements de l\'événement ont été sauvegardé !</string>
 </resources>
